Sending claims was failing for offices using random primary keys sometimes.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@10445 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Main Modules/ContrAccount.cs
@@ -2745,11 +2745,11 @@ namespace OpenDental {
 		}
 
 		private bool CheckClearinghouseDefaults() {
-			if(PrefC.GetInt(PrefName.ClearinghouseDefaultDent)==0) {
+			if(PrefC.GetLong(PrefName.ClearinghouseDefaultDent)==0) {
 				MsgBox.Show(this,"No default dental clearinghouse defined.");
 				return false;
 			}
-			if(PrefC.GetBool(PrefName.ShowFeatureMedicalInsurance) && PrefC.GetInt(PrefName.ClearinghouseDefaultMed)==0) {
+			if(PrefC.GetBool(PrefName.ShowFeatureMedicalInsurance) && PrefC.GetLong(PrefName.ClearinghouseDefaultMed)==0) {
 				MsgBox.Show(this,"No default medical clearinghouse defined.");
 				return false;
 			}
